+++ internal/write/queue_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by MIT
 // license that can be found in the LICENSE file.
 
-package influxdb2
+package write
 
 import (
 	"testing"
@@ -13,7 +13,7 @@ import (
 func TestQueue(t *testing.T) {
 	que := newQueue(2)
 	assert.True(t, que.isEmpty())
-	b := &batch{batch: "batch", retryInterval: 3, retries: 3}
+	b := &Batch{batch: "batch", retryInterval: 3, retries: 3}
 	que.push(b)
 	assert.False(t, que.isEmpty())
 	b2 := que.pop()