+++ options.go
@@ -6,86 +6,73 @@ package influxdb2
 
 import (
 	"crypto/tls"
+	"github.com/influxdata/influxdb-client-go/api/http"
+	"github.com/influxdata/influxdb-client-go/api/write"
 	"time"
 )
 
 // Options holds configuration properties for communicating with InfluxDB server
 type Options struct {
-	// Maximum number of points sent to server in single request. Default 5000
-	batchSize uint
-	// Interval, in ms, in which is buffer flushed if it has not been already written (by reaching batch size) . Default 1000ms
-	flushInterval uint
-	// Default retry interval in ms, if not sent by server. Default 1000ms
-	retryInterval uint
-	// Maximum count of retry attempts of failed writes
-	maxRetries uint
-	// Maximum number of points to keep for retry. Should be multiple of BatchSize. Default 10,000
-	retryBufferLimit uint
-	// DebugLevel to filter log messages. Each level mean to log all categories bellow. 0 error, 1 - warning, 2 - info, 3 - debug
+	// LogLevel to filter log messages. Each level mean to log all categories bellow. 0 error, 1 - warning, 2 - info, 3 - debug
 	logLevel uint
-	// Precision to use in writes for timestamp. In unit of duration: time.Nanosecond, time.Microsecond, time.Millisecond, time.Second
-	// Default time.Nanosecond
-	precision time.Duration
-	// Whether to use GZip compression in requests. Default false
-	useGZip bool
-	// TLS configuration for secure connection. Default nil
-	tlsConfig *tls.Config
-	// HTTP request timeout in sec. Default 20
-	httpRequestTimeout uint
+	// Writing options
+	writeOptions *write.Options
+	// Http options
+	httpOptions *http.Options
 }
 
 // BatchSize returns size of batch
 func (o *Options) BatchSize() uint {
-	return o.batchSize
+	return o.WriteOptions().BatchSize()
 }
 
 // SetBatchSize sets number of points sent in single request
 func (o *Options) SetBatchSize(batchSize uint) *Options {
-	o.batchSize = batchSize
+	o.WriteOptions().SetBatchSize(batchSize)
 	return o
 }
 
 // FlushInterval returns flush interval in ms
 func (o *Options) FlushInterval() uint {
-	return o.flushInterval
+	return o.WriteOptions().FlushInterval()
 }
 
 // SetFlushInterval sets flush interval in ms in which is buffer flushed if it has not been already written
 func (o *Options) SetFlushInterval(flushIntervalMs uint) *Options {
-	o.flushInterval = flushIntervalMs
+	o.WriteOptions().SetFlushInterval(flushIntervalMs)
 	return o
 }
 
 // RetryInterval returns the retry interval in ms
 func (o *Options) RetryInterval() uint {
-	return o.retryInterval
+	return o.WriteOptions().RetryInterval()
 }
 
 // SetRetryInterval sets retry interval in ms, which is set if not sent by server
 func (o *Options) SetRetryInterval(retryIntervalMs uint) *Options {
-	o.retryInterval = retryIntervalMs
+	o.WriteOptions().SetRetryInterval(retryIntervalMs)
 	return o
 }
 
 // MaxRetries returns maximum count of retry attempts of failed writes
 func (o *Options) MaxRetries() uint {
-	return o.maxRetries
+	return o.WriteOptions().MaxRetries()
 }
 
 // SetMaxRetries sets maximum count of retry attempts of failed writes
 func (o *Options) SetMaxRetries(maxRetries uint) *Options {
-	o.maxRetries = maxRetries
+	o.WriteOptions().SetMaxRetries(maxRetries)
 	return o
 }
 
 // RetryBufferLimit returns retry buffer limit
 func (o *Options) RetryBufferLimit() uint {
-	return o.retryBufferLimit
+	return o.WriteOptions().RetryBufferLimit()
 }
 
 // SetRetryBufferLimit sets maximum number of points to keep for retry. Should be multiple of BatchSize.
 func (o *Options) SetRetryBufferLimit(retryBufferLimit uint) *Options {
-	o.retryBufferLimit = retryBufferLimit
+	o.WriteOptions().SetRetryBufferLimit(retryBufferLimit)
 	return o
 }
 
@@ -103,49 +90,65 @@ func (o *Options) SetLogLevel(logLevel uint) *Options {
 
 // Precision returns time precision for writes
 func (o *Options) Precision() time.Duration {
-	return o.precision
+	return o.WriteOptions().Precision()
 }
 
 // SetPrecision sets time precision to use in writes for timestamp. In unit of duration: time.Nanosecond, time.Microsecond, time.Millisecond, time.Second
 func (o *Options) SetPrecision(precision time.Duration) *Options {
-	o.precision = precision
+	o.WriteOptions().SetPrecision(precision)
 	return o
 }
 
 // UseGZip returns true if write request are gzip`ed
 func (o *Options) UseGZip() bool {
-	return o.useGZip
+	return o.WriteOptions().UseGZip()
 }
 
 // SetUseGZip specifies whether to use GZip compression in write requests.
 func (o *Options) SetUseGZip(useGZip bool) *Options {
-	o.useGZip = useGZip
+	o.WriteOptions().SetUseGZip(useGZip)
 	return o
 }
 
 // TlsConfig returns TlsConfig
 func (o *Options) TlsConfig() *tls.Config {
-	return o.tlsConfig
+	return o.HttpOptions().TlsConfig()
 }
 
 // SetTlsConfig sets TLS configuration for secure connection
 func (o *Options) SetTlsConfig(tlsConfig *tls.Config) *Options {
-	o.tlsConfig = tlsConfig
+	o.HttpOptions().SetTlsConfig(tlsConfig)
 	return o
 }
 
 // HttpRequestTimeout returns HTTP request timeout
 func (o *Options) HttpRequestTimeout() uint {
-	return o.httpRequestTimeout
+	return o.HttpOptions().HttpRequestTimeout()
 }
 
 // SetHttpRequestTimeout sets HTTP request timeout in sec
 func (o *Options) SetHttpRequestTimeout(httpRequestTimeout uint) *Options {
-	o.httpRequestTimeout = httpRequestTimeout
+	o.HttpOptions().SetHttpRequestTimeout(httpRequestTimeout)
 	return o
+}
+
+// WriteOptions returns write related options
+func (o *Options) WriteOptions() *write.Options {
+	if o.writeOptions == nil {
+		o.writeOptions = write.DefaultOptions()
+	}
+	return o.writeOptions
+}
+
+// HttpOptions returns http related options
+func (o *Options) HttpOptions() *http.Options {
+	if o.httpOptions == nil {
+		o.httpOptions = http.DefaultOptions()
+	}
+	return o.httpOptions
 }
 
 // DefaultOptions returns Options object with default values
 func DefaultOptions() *Options {
-	return &Options{batchSize: 5000, maxRetries: 3, retryInterval: 1000, flushInterval: 1000, precision: time.Nanosecond, useGZip: false, retryBufferLimit: 10000, httpRequestTimeout: 20}
+	return &Options{logLevel: 0, writeOptions: write.DefaultOptions(), httpOptions: http.DefaultOptions()}
 }